+++ telegram_bot/db.py
@@ -3,7 +3,7 @@ from sqlalchemy import create_engine
 from sqlalchemy.sql import text
 from dotenv import load_dotenv
 
-from constants import INTEREST_RATE_QUERY
+from data_scraper.constants import INTEREST_RATE_QUERY
 from utils import is_docker_env
 
 # загружаем переменные среды
